requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
statsmodels>=0.14.0
matplotlib>=3.7.0
plotly>=5.15.0
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0

# Machine Learning
xgboost>=1.7.0
lightgbm>=4.0.0
torch>=2.0.0
pytorch-lightning>=2.0.0

# Time series
pmdarima>=2.0.0
tsfresh>=0.20.0

# Portfolio optimization and risk
cvxpy>=1.3.0
PyPortfolioOpt>=1.5.0
arch>=5.3.0

# Backtesting
vectorbt>=0.25.0

# Performance optimization
numba>=0.57.0

# Web framework for demo
streamlit>=1.25.0
fastapi>=0.100.0
uvicorn>=0.22.0

# Development tools
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional tracking
mlflow>=2.5.0
